moves model class to model.py for consistency
