+++ padertorch/contrib/examples/speech_enhancement/simple_mask_estimator/__init__.py
@@ -1,1 +1,1 @@
-from .train import SimpleMaskEstimator
+from .model import SimpleMaskEstimator
